refactor: reorganize tools into *_core.py + thin-wrapper pattern
Establish a consistent architecture across all tool domains where
pure-Python logic lives in *_core.py modules and LangChain @tool /
MCP @mcp.tool wrappers are thin one-liner delegates.

Phase 1 - Rename and relocate:
- Rename tools/core.py -> tools/ase_core.py for clarity
- Move MACE schemas from parsl_tools.py to schemas/mace_parsl_schema.py
- Add missing __init__.py in mcp/ and schemas/calculators/

Phase 2 - Extract core modules:
- Create cheminformatics_core.py (deduplicate RDKit/PubChem logic
  from 3 files into single smiles_to_3d implementation)
- Create xanes_core.py (extract from xanes_tools.py, delete
  redundant xanes_mcp.py by merging into xanes_mcp_parsl.py)
- Create graspa_core.py (extract from graspa_tools.py)

Phase 3 - Consolidate shared utilities:
- Add extract_output_json_core to ase_core.py (was duplicated 3x)
- Consolidate load_parsl_config into mcp/server_utils.py (was 2x)
- Replace hardcoded element_map dicts in report_tools.py with
  ase.data.chemical_symbols (handles all elements, not just 1-18)

Phase 4 - Cleanup:
- Delete mcp_helper.py shim, update all consumers
- Refactor new_eval MCP example scripts from ~486 to ~88 lines each
- Fix stale import in utils/tool_call_eval.py

Net result: ~2769 lines removed, ~2783 added (new core modules),
with significant deduplication. All 322 existing tests pass.
